+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -333,5 +333,34 @@ End Sub";
         }
 
 
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void ImplicitPublicMember_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo(ByVal arg1 as Integer)
+'Just an inoffensive little comment
+
+End Sub";
+
+            const string expectedCode =
+@"'@Ignore ImplicitPublicMember
+Sub Foo(ByVal arg1 as Integer)
+'Just an inoffensive little comment
+
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new ImplicitPublicMemberInspection(state);
+            var inspectionResults = inspection.GetInspectionResults();
+
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
     }
 }
